pist1

+++ database/migrations/2021_02_16_120318_create_users_table.php
@@ -16,8 +16,8 @@ class CreateUsersTable extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('quantity');
-            $table->decimal('proteins');
+            $table->float('quantity', 8, 2);
+            $table->float('proteins', 8, 2);
             $table->timestamps();
         });
     }
